requirements.txt
lifxlan
numpy
Pillow